\name{deseasonalize-package}
\alias{deseasonalize-package}
\docType{package}

\title{
Optimal deseasonalization for geophysical time series using AR fitting
}

\description{
Deseasonalize daily or monthly time series.
}

\details{
\tabular{ll}{
Package: \tab deseasonalize\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-12-21\cr
License: \tab GPL (>= 2.10)\cr
LazyLoad: \tab yes\cr
LazyData: \tab yes\cr
}

The only function is \code{\link{ds}}. 

For how to use the funciton \code{\link{ds}}, see the examples provided with the datasets
\code{\link{Saugeen}},
\code{\link{ManleyTemp}},
 and
\code{\link{LondonMaxTemp}}.

}

\author{
A. I. McLeod
 <aimcleod@uwo.ca>
}

\references{

K. W. Hipel and A. I. McLeod (1994). 
Time Series Modelling of Water Resources and Environmental Systems.
Elsevier.
\url{http://www.stats.uwo.ca/faculty/aim/1994Book/default.htm}


McLeod, A.I. and Zhang, Y. (2008b).
Improved Subset Autoregression: With R Package.
Journal of Statistical Software.
}


\keyword{ package }

