\name{ManleyTemp}
\alias{ManleyTemp}
\docType{data}
\title{
Mean monthly temperatures in England (F), 1723-1970, 2976 VALUES                         
}
\description{
Mean monthly temperatures in England (deg F), January 1723 to December 1970.
There are 2976 values.                         

}
\usage{data(ManleyTemp)}
\format{
  The format is:
 Time-Series [1:2976] from 1723 to 1971: 1.1 4.4 7.5 8.9 11.7 15 15.3 15.6 13.3 11.1 ...
}


\source{
Manley, G. (1953).
The Mean Temperatures of Central England (1698-1952). 
Quarterly Journal of the Royal Meteorological Society, 79, 242-261.

Manley, G. (1974).
Central England Temperatures: Monthly Means 1659 to 1973.", 
Quarterly J. of the Royal Meteorological Society, 100, 389-405.
}

\references{
K. W. Hipel and A. I. McLeod (1994). 
Time Series Modelling of Water Resources and Environmental Systems.
Elsevier.
\url{http://www.stats.uwo.ca/faculty/aim/1994Book/default.htm}

\url{http://en.wikipedia.org/wiki/Central_England_temperature}
}

\examples{
plot(ManleyTemp)
}
\keyword{datasets}
