\name{SaugeenDay}
\alias{SaugeenDay}
\docType{data}
\title{
Daily flow Saugeen River, 1915/01/01-1979/12/31
}

\description{
Mean daily flow in
cubic meters per second (cumecs)
of the Saugeen River at Walkerton, Jan 1, 1915 to Dec 31, 1979

}

\usage{data(SaugeenDay)}

\format{
  The format is:
 num [1:23741, 1] 11.5 10.8 13.7 13.7 14.4 17 17 17.8 17.8 17 ...
 - attr(*, "dimnames")=List of 2
  ..$ : chr [1:23741] "1915-01-01" "1915-01-02" "1915-01-03" "1915-01-04" ...
  ..$ : chr "flow"
}

\source{
Environment Canada
}

\references{
K. W. Hipel and A. I. McLeod (1994). 
Time Series Modelling of Water Resources and Environmental Systems.
Elsevier.
}

\examples{
str(SaugeenDay)
}

\keyword{datasets}


